refactor(gsd-db): lift reads into db/queries.ts — the read seam (candidate 3)
Phase A, commit 5. Moves the 34 SELECT-only wrappers (+ the VerificationEvidenceRow
type and two private parse helpers they use) out of gsd-db.ts into db/queries.ts,
re-exported by the barrel. Read-only callers (forensics, dashboard, doctor) now
depend on a read seam, not the single-writer surface.

Adds an explicit invariant assertion that db/queries.ts contains no write SQL
(test 1 already forbids it structurally — queries.ts is under db/ but not in
db/writers/ — this states the read-seam intent positively).

gsd-db.ts: 1,853 → 1,484 lines (from 3,644 at the start of Phase A).

Verified: typecheck:extensions clean; single-writer-invariant (incl. new
read-only assertion) + db-authority-regression + completion-hierarchy-guards +
context-store-decisions + forensics green (46 tests).
